making progress for broadcast project

+++ template_java/src/main/java/cs451/Host.java
@@ -50,8 +50,8 @@ public class Host {
         return true;
     }
 
-    public int getId() {
-        return id;
+    public short getId() {
+        return (short)id;
     }
 
     public String getIp() {
